renamed api handling lambda and removed unused code
