+++ cloudformation/cloudformation.yaml
@@ -63,13 +63,13 @@ Resources:
                            Action:
                            - s3:ListBucket
                            Resource: arn:aws:s3:::fulfilment-output-test
-  FulfilmentTriggerLambda:
+  SalesforceUploaderLambda:
     Type: "AWS::Lambda::Function"
     Properties:
       FunctionName:
-          !Sub fulfilment_trigger-${Stage}
-      Description: "Trigger fulfilment step functions"
-      Handler: "fulfilmentTrigger.handler"
+          !Sub salesforce_uploader-${Stage}
+      Description: "upload fulfilment files to salesforce"
+      Handler: "salesforce_uploader.handler"
       Role: !GetAtt [ FulfilmentWorkersLambdaRole, Arn ]
       Code:
         S3Bucket: fulfilment-lambdas-dist
@@ -87,7 +87,7 @@ Resources:
     Properties:
       FunctionName:
           !Sub zuora_fulfilment_querier-${Stage}
-      Description: "Trigger zuora export"
+      Description: "Fulzuora export"
       Handler: "querier.handler"
       Role: !GetAtt [ FulfilmentWorkersLambdaRole, Arn ]
       Code:
@@ -250,14 +250,14 @@ Resources:
           Type: AWS::Lambda::Permission
           Properties:
               Action: lambda:invokeFunction
-              FunctionName: !Sub fulfilment_trigger-${Stage}
+              FunctionName: !Sub salesforce_uploader-${Stage}
               Principal: apigateway.amazonaws.com
-          DependsOn: FulfilmentTriggerLambda
+          DependsOn: SalesforceUploaderLambda
 
   FulfilmentAPI:
       Type: "AWS::ApiGateway::RestApi"
       Properties:
-          Description: triggers fulfilment file generation
+          Description: upload fulfilment files to salesforce
           Name: !Sub fulfilment-api-${Stage}
 
   FulfilmentProxyResource:
@@ -278,10 +278,10 @@ Resources:
           Integration:
             Type: AWS_PROXY
             IntegrationHttpMethod: POST
-            Uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${FulfilmentTriggerLambda.Arn}/invocations
+            Uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SalesforceUploaderLambda.Arn}/invocations
       DependsOn:
       - FulfilmentAPI
-      - FulfilmentTriggerLambda
+      - SalesforceUploaderLambda
       - FulfilmentProxyResource
 
   FulfilmentAPIStage:
