+++ test/upload.json
@@ -1,3 +1,6 @@
 {
-  "path":"test.txt"
+	"headers": {
+		"apiToken": "[SECRET TOKEN HERE]"
+	},
+	"body": "{\n\t\"date\":\"2017-07-12\",\n\t\"amount\":1\n}"
 }
